Move filter utils into a new file
